포스트 삭제 시 코멘트와 좋아요도 삭제

+++ src/main/java/com/fastcampus/sns/controller/PostController.java
@@ -54,8 +54,8 @@ public class PostController {
     }
 
     @GetMapping("/{postId}/likes")
-    public Response<Integer> likeCount(@PathVariable Integer postId) {
-        int count = postService.likeCount(postId);
+    public Response<Long> likeCount(@PathVariable Integer postId) {
+        long count = postService.likeCount(postId);
         return Response.success(count);
     }
